requirements.txt
tensorflow>=2.10.0
rasterio>=1.3.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.64.0